hi lol

+++ src/common/mod.rs
@@ -2,4 +2,4 @@ pub mod cue;
 pub mod id;
 pub mod workspace;
 
-pub type CueId = u64;
+pub type CueId = u32;
